Sys0 ownable (#52)
* stash new structure; with ownable bug

* compiling

* prettier everything

+++ contracts/proto0/OracleP0.sol
@@ -1,18 +1,20 @@
 // SPDX-License-Identifier: BlueOak-1.0.0
 pragma solidity 0.8.4;
 
-import "@openzeppelin/contracts/access/Ownable.sol";
+import "../Ownable.sol"; // temporary
+// import "@openzeppelin/contracts/access/Ownable.sol";
+
 import "../libraries/CommonErrors.sol";
 import "./interfaces/ICollateral.sol";
 
 contract Oracle is Ownable {
-    mapping(address => uint) prices;
+    mapping(address => uint256) prices;
 
-    function setPrice(address token, uint amount) public onlyOwner{
+    function setPrice(address token, uint256 amount) public onlyOwner {
         prices[token] = amount;
     }
 
-    function getPrice(address token) public view returns (uint) {
+    function getPrice(address token) public view returns (uint256) {
         if (prices[token] == 0) {
             revert CommonErrors.PriceNotFound();
         }
