+++ contracts/proto0/collateral/CTokenCollateral.sol
@@ -6,13 +6,15 @@ import "./Collateral.sol";
 
 // https://github.com/compound-finance/compound-protocol/blob/master/contracts/CToken.sol
 interface ICToken {
-    function exchangeRateCurrent() external returns (uint);
+    function exchangeRateCurrent() external returns (uint256);
 }
 
-
 contract CTokenCollateral is Collateral {
-
-    constructor(address erc20_, uint256 quantity_, uint8 decimals) Collateral(erc20_, quantity_, decimals) {}
+    constructor(
+        address erc20_,
+        uint256 quantity_,
+        uint8 decimals
+    ) Collateral(erc20_, quantity_, decimals) {}
 
     function getRedemptionRate() external override returns (uint256) {
         return ICToken(_erc20).exchangeRateCurrent();
